requirements.txt
numpy==1.19.0
Pillow==7.2.0
grpcio==1.30.0
